reach-skyline-backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Cors==4.0.1
flask-jwt-extended>=4.6.0
PyMySQL==1.1.0
python-dotenv==1.0.1
Werkzeug==3.0.3
alembic==1.17.2
google-api-python-client==2.128.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
flask-socketio==5.3.6
python-socketio==5.11.0
python-engineio==4.9.0
pip-system-certs  # In case of SSL issues in some environments

